document management was modified

+++ CollectionManagement.py
@@ -5,6 +5,7 @@ import azure.cosmos.errors as errors
 import shared.config as cfg
 from DBManagement import DatabaseManagement
 
+
 # ----------------------------------------------------------------------------------------------------------
 # Prerequistes - 
 # 
@@ -16,22 +17,22 @@ from DBManagement import DatabaseManagement
 # ----------------------------------------------------------------------------------------------------------
 # Sample - demonstrates the basic CRUD operations on a Collection resource for Azure Cosmos
 # 
-# 1. Query for Collection
+#  Query for Collection
 #  
-# 2. Create Collection
-#    2.1 - Basic Create
-#    2.2 - Create collection with custom IndexPolicy
-#    2.3 - Create collection with offer throughput set
+#  Create Collection
+#    
+#     Create collection with custom IndexPolicy 
+#     Create collection with offer throughput set
 #
-# 3. Manage Collection Offer Throughput
-#    3.1 - Get Collection performance tier
-#    3.2 - Change performance tier
+#  Manage Collection Offer Throughput
+#    Get Collection performance tier
+#    Change performance tier
 #
-# 4. Get a Collection by its Id property
+#  Get a Collection by its Id property
 #
-# 5. List all Collection resources in a Database
+#  List all Collection resources in a Database
 #
-# 6. Delete Collection
+#  Delete Collection
 # ----------------------------------------------------------------------------------------------------------
 # Note - 
 # 
@@ -45,30 +46,16 @@ MASTER_KEY = cfg.settings['master_key']
 DATABASE_ID = cfg.settings['database_id']
 COLLECTION_ID = cfg.settings['collection_id']
 
-database_link = 'dbs/' + DATABASE_ID
-collections_link ='dbs/' + DATABASE_ID + '/colls'
-
-class IDisposable:
-    """ A context manager to automatically close an object with a close method
-    in a with statement. """
-
-    def __init__(self, obj):
-        self.obj = obj
-
-    def __enter__(self):
-        return self.obj # bound to target
-
-    def __exit__(self, exception_type, exception_val, trace):
-        # extra cleanup in here
-        self = None
+# database_link = 'dbs/' + DATABASE_ID
+# collections_link ='dbs/' + DATABASE_ID + '/colls'
 
 class CollectionManagement:
     @staticmethod
-    def find_Container(client, id):
-        print('1. Query for specific Collection')
-
+    def find_Container(client, db, id):
+        print('Query for specific Collection')
+        db_link = 'dbs/' + db
         collections = list(client.QueryContainers(
-            database_link,
+            db_link,
             {
                 "query": "SELECT * FROM r WHERE r.id=@id",
                 "parameters": [
@@ -79,23 +66,14 @@ class CollectionManagement:
 
         if len(collections) > 0:
             print('Collection with id \'{0}\' was found'.format(id))
+            return True
         else:
-            print('No collection with id \'{0}\' was found'. format(id))
-
-    # @staticmethod
-    # def list_all_Containers(client):
-    #     print('2. Query for Collections')
-
-    #     collections = list(client.QueryContainers(collection_link)
-
-    #     if len(collections) > 0:
-    #         for c in collections: 
-    #             print('Collection with id \'{0}\' was found'.format(c))
-    #     else:
-    #         print('No collection was found')
-
+            print('No collection with id \'{0}\' was found'.format(id))
+            return False
+    
+    
     @staticmethod
-    def create_Container(client, id):
+    def create_Container(client, db, id):
         """ The most basic Create of collection will create a collection with 400 RUs throughput and default automatic indexing policy          
          Our code will create a collection with custom index policy, custom offer throughput and unique keys     
          """
@@ -116,8 +94,11 @@ class CollectionManagement:
                     }]
                 }
             }
-
-            collection = client.CreateContainer(database_link, coll)
+            if DatabaseManagement.find_database(client, db):
+                db_link = 'dbs/' + db
+            else:
+                return
+            collection = client.CreateContainer(db_link, coll)
             unique_key_paths = collection['uniqueKeyPolicy']['uniqueKeys'][0]['paths']
             print('Collection with id \'{0}\' created'.format(collection['id']))
             print('IndexPolicy Mode - \'{0}\''.format(collection['indexingPolicy']['indexingMode']))
@@ -131,11 +112,11 @@ class CollectionManagement:
                print('A collection with id \'{0}\' already exists'.format(collection['id']))
             else: 
                 raise errors.HTTPFailure(e.status_code) 
-
+        
 
     @staticmethod
-    def manage_offer_throughput(client, id):
-        print("\n3.1 Get Collection Performance tier")
+    def manage_offer_throughput(client, db, id):
+        print("\nGet Collection Performance tier")
         
         #A Collection's Offer Throughput determines the performance throughput of a collection. 
         #A Collection is loosely coupled to Offer through the Offer's offerResourceId
@@ -144,7 +125,7 @@ class CollectionManagement:
         
         try:
             # read the collection, so we can get its _self
-            collection_link = database_link + '/colls/{0}'.format(id)
+            collection_link = 'dbs/'+ db +'/colls/{0}'.format(id)
             collection = client.ReadContainer(collection_link)
             # print('collection name \'{0}\''.format(collection))
 
@@ -159,7 +140,7 @@ class CollectionManagement:
             else: 
                 raise errors.HTTPFailure(e.status_code)
 
-        print("\n3.2 Change Offer Throughput of Collection")
+        print("\nChange Offer Throughput of Collection")
                            
         #The Offer Throughput of a collection controls the throughput allocated to the Collection
         #To increase (or decrease) the throughput of any Collection you need to adjust the Offer.content.offerThroughput
@@ -172,49 +153,58 @@ class CollectionManagement:
         print('Replaced Offer. Offer Throughput is now \'{0}\''.format(offer['content']['offerThroughput']))
                                 
     @staticmethod
-    def read_Container(client, id):
-        print("\n Get a Collection by id")
+    def read_Container(client, db, id):
+        print("\nGet a Collection by id")
 
         try:
             # All Azure Cosmos resources are addressable via a link
             # This link is constructed from a combination of resource hierachy and 
             # the resource id. 
             # Eg. The link for collection with an id of Bar in database Foo would be dbs/Foo/colls/Bar
-            collection_link = database_link + '/colls/{0}'.format(id)
-
+            db_link = 'dbs/' + db
+            collection_link = db_link + '/colls/{0}'.format(id)
             collection = client.ReadContainer(collection_link)
             print('Collection with id \'{0}\' was found, it\'s _self is {1}'.format(collection['id'], collection['_self']))
-
+            
         except errors.HTTPFailure as e:
             if e.status_code == 404:
                print('A collection with id \'{0}\' does not exist'.format(id))
             else: 
                 raise errors.HTTPFailure(e.status_code)    
-    
+        return True
+
     @staticmethod
     def list_Containers(client, db):
-
-        if  DatabaseManagement.find_database(client, db):
-            print("\nList all collections in database \'{0}\'".format(db))               
-            db_link = 'dbs/' + db
-            collections = list(client.ReadContainers(db_link))            
-            if not collections:
-                print("\'{0}\' has no collections".format(db))
-                return
-            for collection in collections:
-                print(collection['id'])          
-        else:
-            print("\'{0}\' not found".format(db))
-            
+        try:
+            if  DatabaseManagement.find_database(client, db):
+                print("\nList all collections in database \'{0}\'".format(db))               
+                db_link = 'dbs/' + db
+                collections = list(client.ReadContainers(db_link))            
+                if not collections:
+                    print("\'{0}\' has no collections".format(db))
+                    return
+                for collection in collections:
+                    print(collection['id'])          
+            else:
+                print("\'{0}\' not found".format(db))
+        except errors.HTTPFailure as e:
+            if e.status_code == 404:
+               print("{0} has no collections".format(db))
+            else: 
+                raise errors.HTTPFailure(e.status_code)   
     @staticmethod
-    def delete_Container(client, id):
-        print("\n6. Delete Collection")
+    def delete_Container(client, db, id):
+        print("\nDelete Collection")
         
         try:
-           collection_link = database_link + '/colls/{0}'.format(id)
-           client.DeleteContainer(collection_link)
+            if DatabaseManagement.find_database(client, db) and CollectionManagement.find_Container(client, id):
+                db_link = 'dbs/' + db
+                collection_link = db_link + '/colls/{0}'.format(id)
+                client.DeleteContainer(collection_link)
+                print('Collection with id \'{0}\' was deleted'.format(id))
+            else:
+                return
 
-           print('Collection with id \'{0}\' was deleted'.format(id))
 
         except errors.HTTPFailure as e:
             if e.status_code == 404:
